feat(refactor): seperate db into multiple files, removed db test
