+++ backend/package.json
@@ -7,7 +7,6 @@
     "build": "tsc",
     "start": "node dist/server.js",
     "setup-db": "ts-node setup-database.ts",
-    "test-rds": "ts-node test-rds-connection.ts",
     "log-db": "ts-node log-db.ts"
   },
   "keywords": [],
